More WIPs - Assigning network data causes new GUIDs to be created for Attributes. This is not desireable as new GUIDS cannot be linked to data in that network and the network and its scenarios should be independent from the DATA aggregation (barring any Attribute changes).
- The data assignment process will need to check to see if each attribute it finds already exists (via new Guid in the meta data file) in the meta data file.

If the GUIDs of an attribute from the database match a GUID in the file, then we do not create a new GUID.

If an attribute in the database cannot be found in the file (no matching GUID) then we skip aggregation for that attribute.

If an attribute in the meta data file cannot be found in the database, we create a new GUID and aggregate the data.

This should ensure that ANY changes made to existing attributes in a network, keep the network and its scenarios running correctly.